+++ src/main/java/shadows/apotheosis/mixin/ArrowItemMixin.java
@@ -9,13 +9,14 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ArrowItem;
 import net.minecraft.world.item.ItemStack;
 import shadows.apotheosis.Apoth;
+import shadows.apotheosis.Apotheosis;
 
 @Mixin(ArrowItem.class)
 public class ArrowItemMixin {
 
 	@Inject(method = "isInfinite", at = @At(value = "RETURN"), remap = false, cancellable = true)
 	public void apoth_isInfinite(ItemStack stack, ItemStack bow, Player player, CallbackInfoReturnable<Boolean> ci) {
-		if (!ci.getReturnValueZ()) {
+		if (!ci.getReturnValueZ() && Apotheosis.enableEnch) {
 			ci.setReturnValue(Apoth.Enchantments.ENDLESS_QUIVER.isTrulyInfinite(stack, bow, player));
 		}
 	}